Refactors code
- Replaces occurrences of deprecated cons with new cons
- Moves common methods to MqttAPIHelper class
